fix: call resolver helper inside spell handler
